Documents/AI/Driving_Forward_Simulation/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0

# Backbone
timm==0.9.12
einops==0.7.0

# Dataset
nuscenes-devkit>=1.1.11
pyquaternion

# Image processing
opencv-python==4.9.0.80
Pillow>=10.0.0

# Metrics / losses
lpips==0.1.4
scikit-image==0.22.0

# Config & logging
PyYAML==6.0.1
tensorboard==2.15.2
tqdm==4.66.2
python-dotenv==1.0.0

# Visualisation
matplotlib>=3.8.0
open3d>=0.18.0   # optional, for 3-D point cloud vis
